[ENH] pabi_utils, feature to report as CSV

+++ account_budget_activity/report/product_activity.py
@@ -24,12 +24,12 @@ class ProductActivity(models.Model):
         select id, temp_product_id, temp_activity_id, name
         from ((
             select 1000000 + p.id id,
-            p.id temp_product_id, null temp_activity_id,
+            p.id temp_product_id, null::int temp_activity_id,
             p.name_template as name
             from product_product p)
         union all (
             select 2000000 + a.id,
-            null temp_product_id, a.id temp_activity_id,
+            null::int temp_product_id, a.id temp_activity_id,
             a.name as name
             from account_activity a
         )) product_activity
